move common types
