+++ fastcrypto-tbls/src/types.rs
@@ -2,9 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::polynomial::{Eval, PublicPoly};
-use crate::{ecies_v0, tbls};
+use crate::tbls;
 use fastcrypto::error::{FastCryptoError, FastCryptoResult};
-use fastcrypto::groups::ristretto255::RistrettoPoint;
 use fastcrypto::groups::{bls12381, GroupElement, HashToGroupElement, Pairing};
 use serde::{Deserialize, Serialize};
 use std::num::NonZeroU16;
@@ -84,10 +83,3 @@ impl<A> UnindexedValues<A> {
         Ok(values)
     }
 }
-
-/// ECIES related types with Ristretto points.
-///
-pub type PrivateEciesKey = ecies_v0::PrivateKey<RistrettoPoint>;
-pub type PublicEciesKey = ecies_v0::PublicKey<RistrettoPoint>;
-pub type EciesEncryption = ecies_v0::Encryption<RistrettoPoint>;
-pub type RecoveryPackage = ecies_v0::RecoveryPackage<RistrettoPoint>;
